separate zstd registry
